Splitting all eaders into .h and .cpp

+++ Sources/VisualComponent.h
@@ -38,7 +38,7 @@ public:
 			vkCmdBindVertexBuffers(commandBuffer, 0, 1, &m_ModelData->meshes[i].vertexBufffer, offsets);
 
 			vkCmdBindIndexBuffer(commandBuffer, m_ModelData->meshes[i].indexBuffer, 0, VK_INDEX_TYPE_UINT32);
-			vkCmdDrawIndexed(commandBuffer, m_ModelData->meshes[i].indexBufferSize, 1, 0, 0, 0);
+			vkCmdDrawIndexed(commandBuffer, static_cast<uint32_t>(m_ModelData->meshes[i].indexBufferSize), 1, 0, 0, 0);
 		}
 	}
